requirements.txt
fastapi
uvicorn
requests
pillow
python-dotenv
tflite-runtime
mangum
